# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: MindTab API
  description: API for MindTab — task tracking and micro-notes platform.
  version: 0.1.0
  contact:
    name: MindTab
    url: https://mindtab.in
servers:
  - url: /
    description: API v1
security:
  - bearerAuth: []
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [Projects]
      summary: List projects with nested tasks
      parameters:
        - name: includeArchived
          in: query
          schema:
            type: boolean
            default: false
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProjectStatus"
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createProject
      tags: [Projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectCreate"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [Projects]
      summary: Get a project with nested tasks
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProject
      tags: [Projects]
      summary: Update a project
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProject
      tags: [Projects]
      summary: Soft-delete a project (cascades to tasks and notes)
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Project deleted
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /projects/{id}/archive:
    post:
      operationId: archiveProject
      tags: [Projects]
      summary: Archive a project and its contents
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Archived project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Activity ───────────────────────────────────────────────────────────
...
  /search/tasks:
    get:
      operationId: searchTasks
      tags: [Search]
      summary: Search tasks by title (max 5 results)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Matching tasks
          content:
            application/json:
              schema:
                type: array
                maxItems: 5
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /search/habits:
    get:
      operationId: searchHabits
      tags: [Search]
      summary: Search habits by title (max 5 results)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Matching habits
          content:
            application/json:
              schema:
                type: array
                maxItems: 5
                items:
                  $ref: "#/components/schemas/Habit"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /search/notes:
    get:
      operationId: searchNotes
      tags: [Search]
      summary: Search notes by title (max 5 results)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Matching notes
          content:
            application/json:
              schema:
                type: array
                maxItems: 5
                items:
                  $ref: "#/components/schemas/Note"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── Sync ───────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Parameters ───────────────────────────────────────────────────────
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  # ─── Reusable Responses ──────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ─── Schemas ──────────────────────────────────────────────────────────
  schemas:
...
    TaskStatus:
      type: string
      enum: [pending, in_progress, completed, archived]
    TaskPriority:
      type: string
      enum: [priority_1, priority_2, priority_3, priority_4]
    TaskImpact:
      type: string
      enum: [low, medium, high]
    HabitFrequency:
      type: string
      enum: [daily, weekly]
...
    NoteType:
      type: string
      enum: [article, book, video, podcast, website]
    ProjectStatus:
      type: string
      enum: [active, paused, completed, archived]

    # ── Error ──────────────────────────────────────────────────────────
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
        code:
          type: string
          description: Machine-readable error code

    # ── Auth ───────────────────────────────────────────────────────────
...
    Task:
      type: object
      required: [id, status, priority, impact, position, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        priority:
          $ref: "#/components/schemas/TaskPriority"
        impact:
          $ref: "#/components/schemas/TaskImpact"
        position:
          type: integer
        projectId:
          type: string
          format: uuid
          nullable: true
        project:
          $ref: "#/components/schemas/ProjectSummary"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
...
    Habit:
      type: object
      required: [id, frequency, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        frequency:
          $ref: "#/components/schemas/HabitFrequency"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
...
    Note:
      type: object
      required: [id, title, content, type, source, createdAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        type:
          $ref: "#/components/schemas/NoteType"
        source:
          type: string
        projectId:
          type: string
          format: uuid
          nullable: true
        project:
          $ref: "#/components/schemas/ProjectSummary"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        deletedAt:
          type: string
          format: date-time
          nullable: true
        archivedAt:
          type: string
          format: date-time
          nullable: true
...
    ProjectSummary:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ProjectStatus"
    Project:
      type: object
      required: [id, status, startDate, createdBy, lastUpdatedBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ProjectStatus"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
        createdBy:
          type: string
          maxLength: 255
        lastUpdatedBy:
          type: string
          maxLength: 255
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/Task"
    ProjectCreate:
      type: object
      required: [name, startDate]
      properties:
        name:
          type: string
          maxLength: 256
        description:
          type: string
        status:
          $ref: "#/components/schemas/ProjectStatus"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          maxLength: 256
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ProjectStatus"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
          nullable: true
...